requirements.txt
# LangChain 1.2
langchain>=1.2.0,<1.3.0
langchain-community>=0.3.0
langchain-deepseek>=1.0.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
aiosqlite>=0.20.0
PyJWT>=2.8.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7.4

# 环境变量
python-dotenv>=1.0.0

# 将 LCEL 链暴露为 HTTP API（与笔记 08 配套）
langserve>=0.3.0
